feature: new route for getting relevant cards
